requirements.txt
certifi>=2022.12.7
charset-normalizer>=3.0.1
decorator>=4.4.2
idna>=3.4
imageio>=2.25.1
imageio-ffmpeg>=0.4.8
moviepy>=1.0.3
opencv-python>=4.7.0.68
Pillow>=9.4.0
proglog>=0.1.10
requests>=2.28.2
tqdm>=4.64.1
urllib3>=1.26.14
numpy>=1.19.1
git+https://github.com/kevin-tofu/coco_formatter.git@main